+++ package.json
@@ -1,28 +1,29 @@
 {
-  "name": "jordanpb.com",
-  "version": "1.0.0",
-  "description": "Beginning of my Website",
-  "main": "index.js",
+  "name": "learn-starter",
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
-    "dev": "next",
-    "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "dev:client": "next src/client",
+    "dev:server": "babel src/server --out-dir dist/server --source-maps --watch",
+    "dev": "npm run dev:client & npm run dev:server",
+    "build:client": "next build src/client",
+    "build:server": "babel src/server --out-dir dist/server --source-maps",
+    "build": "npm run build:client && npm run build:server",
+    "preserve": "npm run build",
+    "serve": "cross-env NODE_ENV=production firebase serve --only functions,hosting",
+    "predeploy": "rimraf dist/ && npm run build",
+    "deploy": "cross-env NODE_ENV=production firebase deploy --only functions,hosting"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/JordanPat/JordansSite.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/JordanPat/JordansSite/issues"
-  },
-  "homepage": "https://github.com/JordanPat/JordansSite#readme",
   "dependencies": {
-    "next": "^9.2.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "next": "9.3.5",
+    "react": "16.13.1",
+    "react-dom": "16.13.1"
   },
-  "devDependencies": {}
+
+  "main": "dist/server/index.js",
+  "engines": {
+      "node": "10"
+  }
+
 }
