+++ src/client/next.config.js
@@ -1,0 +1,1 @@
+distDir: "../../dist/client"
